Added doctor functionalities

+++ src/components/secretary/patients-table/patients-table.component.tsx
@@ -1,17 +1,22 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Table, Button } from "antd";
 import { PatientInfo } from "../../../types/Interfaces";
 import { ColumnsType } from "antd/lib/table";
 import { Confirmation, notify } from "../../global/alerts/alerts.component";
-import { useAppDispatch } from "../../../redux/hooks";
+import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
 import { deletePatient } from "../../../redux/patients/patients.slice";
+import { useHistory, useParams } from "react-router-dom";
 
 interface Props {
 	data: PatientInfo[];
+	isAppointment?: boolean;
 }
 
-const PatientTable: React.FC<Props> = ({ data }) => {
+const PatientTable: React.FC<Props> = ({ data, isAppointment = false }) => {
 	const dispatch = useAppDispatch();
+	const history = useHistory();
+	const { idParam } = useParams<{ idParam: string }>();
+	const currentUserRole = useAppSelector((state) => state?.auth?.user?.role);
 	const columns: ColumnsType<PatientInfo> = [
 		{
 			title: "Full Name",
@@ -65,35 +70,42 @@ const PatientTable: React.FC<Props> = ({ data }) => {
 			dataIndex: "occupation",
 			key: "occupation",
 		},
-		{
-			title: "Action",
-			render: (row): JSX.Element => {
-				return (
-					<div className="flex">
-						<Confirmation
-							confirmFn={(): void => {
-								dispatch(
-									deletePatient(row._id, () => {
-										notify(
-											"Patient deleted!",
-											"success"
-										);
-									})
-								);
-							}}
-							title="Click ok to delete this patient."
-						>
-							<Button size="small" danger>
-								Delete
-							</Button>
-						</Confirmation>
-					</div>
-				);
-			},
-		},
 	];
 
-	return <Table<PatientInfo> dataSource={data} columns={columns} />;
+	useEffect(() => {
+		if (currentUserRole !== "doctor") {
+			columns.push({
+				title: "Action",
+				render: (row): JSX.Element => {
+					return (
+						<div className="flex">
+							<Confirmation
+								confirmFn={(): void => {
+									dispatch(
+										deletePatient(row._id, () => {
+											notify(
+												"Patient deleted!",
+												"success"
+											);
+										})
+									);
+								}}
+								title="Click ok to delete this patient."
+							>
+								<Button size="small" danger>
+									Delete
+								</Button>
+							</Confirmation>
+						</div>
+					);
+				},
+			});
+		}
+	}, []);
+
+	return (
+		<Table<PatientInfo> rowKey="_id" dataSource={data} columns={columns} />
+	);
 };
 
 export default PatientTable;
